requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
google-genai>=1.0.0
python-dotenv>=1.0.0
rich>=13.7.0
